% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.gtv.R
\name{cv.gtv}
\alias{cv.gtv}
\title{Cross-validation for gtv}
\usage{
cv.gtv(X, y, Sigma, family = "gaussian", nlambdas = c(5, 5, 30),
  nfolds = 5, metric = "l2", lambdas.seq = NULL, parallel = FALSE,
  nCores = detectCores() - 1)
}
\description{
Does k-fold cross-validation for gtv
}
