% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleGTV.R
\docType{data}
\name{exampleGTV}
\alias{exampleGTV}
\title{Create Example Data Set}
\format{An object of class \code{list} of length 3.}
\usage{
exampleGTV
}
\description{
a block complete graph G that has K connected components and 
each connected component is a complete graph with p/K nodes
}
\keyword{datasets}
