% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtv.R
\name{gtv}
\alias{gtv}
\title{Solve regression problem with graph-based regularization}
\usage{
gtv(X, y, Sigma, lam_TV, lam_S, lam_1 = NULL, family = "gaussian",
  nlambda = 30, lambda.min.ratio = ifelse(nrow(X) < ncol(X), 0.01,
  1e-04), weights = NULL, eps = 1e-04, init = NULL, maxit = 1000,
  Bt = NULL)
}
\description{
Solve regression problem with graph-based regularization
}
